@@ -20,25 +20,29 @@ import static org.wordpress.android.support.WPSupportUtils.populateTextField;
 import static org.wordpress.android.support.WPSupportUtils.waitForElementToBeDisplayed;
 
 public class SignupFlow {
-    public SignupFlow chooseSignupWithEmail() {
-        clickOn(onView(withId(R.id.create_site_button)));
-        clickOn(onView(withId(R.id.signup_email)));
-
+    public SignupFlow chooseContinueWithWpCom() {
+        // Login Prologue – We want to Continue with WordPress.com, not a site address
+        // See LoginPrologueFragment
+        clickOn(R.id.first_button);
         return this;
     }
 
     public SignupFlow enterEmail(String email) {
         // Email file = id/input
         populateTextField(onView(withId(R.id.input)), email);
-        clickOn(onView(withId(R.id.primary_button)));
-
-        // Should See Open Mail button
-        waitForElementToBeDisplayed(R.id.signup_magic_link_button);
-
+        clickOn(onView(withId(R.id.login_continue_button)));
         return this;
     }
 
     public SignupFlow openMagicLink(ActivityTestRule<LoginMagicLinkInterceptActivity> magicLinkActivityTestRule) {
+        // Receive Magic Link – Choose "Send link by email"
+        // See SignupConfirmationFragment
+        clickOn(R.id.signup_confirmation_button);
+
+        // Should see "Check email" button
+        // See SignupMagicLinkFragment
+        waitForElementToBeDisplayed(R.id.signup_magic_link_button);
+
         // Follow the magic link to continue login
         // Intent is invoked directly rather than through a browser as WireMock is unavailable once in the background
         Intent intent = new Intent(
@@ -53,7 +57,7 @@ public class SignupFlow {
 
     public SignupFlow checkEpilogue(String displayName, String username) {
         // Check Epilogue data
-        ViewInteraction emailHeaderView = onView(withId(R.id.signup_epilogue_header_email));
+        ViewInteraction emailHeaderView = onView(withId(R.id.login_epilogue_header_subtitle));
         waitForElementToBeDisplayed(emailHeaderView);
 
         ViewInteraction displayNameField = onView(allOf(withId(R.id.input), withText(displayName)));
