@@ -10,70 +10,75 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import org.wordpress.android.R;
 import org.wordpress.android.fluxc.model.AccountModel;
-import org.wordpress.android.util.GravatarUtils;
+import org.wordpress.android.fluxc.model.SiteModel;
+import org.wordpress.android.login.util.AvatarHelper.FakeGravatarUtils;
 import org.wordpress.android.util.StringUtils;
 import org.wordpress.android.util.image.ImageManager;
-import org.wordpress.android.util.image.ImageType;
+
+import static org.wordpress.android.login.util.AvatarHelper.FakeGravatarUtils.DefaultImage.STATUS_404;
+import static org.wordpress.android.util.image.ImageType.AVATAR_WITHOUT_BACKGROUND;
 
 /**
  * ViewHolder for a RecyclerView header used on screens shown after a user logs in.
  */
-public class LoginHeaderViewHolder extends RecyclerView.ViewHolder {
-    private final View mLoggedInAsHeading;
-    private final View mUserDetailsCard;
+class LoginHeaderViewHolder extends RecyclerView.ViewHolder {
     private final ImageView mAvatarImageView;
     private final TextView mDisplayNameTextView;
     private final TextView mUsernameTextView;
     private final TextView mMySitesHeadingTextView;
 
-    public LoginHeaderViewHolder(View view) {
+    LoginHeaderViewHolder(View view) {
         super(view);
-        mLoggedInAsHeading = view.findViewById(R.id.logged_in_as_heading);
-        mUserDetailsCard = view.findViewById(R.id.user_details_card);
-        mAvatarImageView = view.findViewById(R.id.avatar);
-        mDisplayNameTextView = view.findViewById(R.id.display_name);
-        mUsernameTextView = view.findViewById(R.id.username);
-        mMySitesHeadingTextView = view.findViewById(R.id.my_sites_heading);
+        mAvatarImageView = view.findViewById(R.id.login_epilogue_header_avatar);
+        mDisplayNameTextView = view.findViewById(R.id.login_epilogue_header_title);
+        mUsernameTextView = view.findViewById(R.id.login_epilogue_header_subtitle);
+        mMySitesHeadingTextView = view.findViewById(R.id.login_epilogue_header_sites_subheader);
     }
 
-    public void updateLoggedInAsHeading(Context context, ImageManager imageManager, boolean isAfterLogin,
-                                        AccountModel defaultAccount) {
-        updateLoggedInAsHeading(context, imageManager, isAfterLogin, constructGravatarUrl(context, defaultAccount),
-                                defaultAccount.getUserName(), defaultAccount.getDisplayName());
+    void updateLoggedInAsHeading(Context context, ImageManager imageManager, AccountModel account) {
+        final String avatarUrl = constructGravatarUrl(context, account);
+        final String username = account.getUserName();
+        final String displayName = account.getDisplayName();
+        updateLoggedInAsHeading(imageManager, avatarUrl, username, displayName);
     }
 
-    public void updateLoggedInAsHeading(Context context, ImageManager imageManager, boolean isAfterLogin,
-                                        String avatarUrl, String username, String displayName) {
-        if (isAfterLogin) {
-            mLoggedInAsHeading.setVisibility(View.VISIBLE);
-            mUserDetailsCard.setVisibility(View.VISIBLE);
-            imageManager.loadIntoCircle(mAvatarImageView, ImageType.AVATAR_WITHOUT_BACKGROUND,
-                    StringUtils.notNullStr(avatarUrl));
-            mUsernameTextView.setText(context.getString(R.string.login_username_at, username));
+    void updateLoggedInAsHeading(Context context, ImageManager imageManager, SiteModel site) {
+        final String avatarUrl = constructGravatarUrl(context, site);
+        final String username = site.getUsername();
+        final String displayName = site.getDisplayName();
+        updateLoggedInAsHeading(imageManager, avatarUrl, username, displayName);
+    }
 
-            if (!TextUtils.isEmpty(displayName)) {
-                mDisplayNameTextView.setText(displayName);
-            } else {
-                mDisplayNameTextView.setText(username);
-            }
+    void updateLoggedInAsHeading(ImageManager imageManager, String avatarUrl, String username, String displayName) {
+        imageManager.loadIntoCircle(mAvatarImageView, AVATAR_WITHOUT_BACKGROUND, StringUtils.notNullStr(avatarUrl));
+
+        mUsernameTextView.setText(username);
+
+        if (!TextUtils.isEmpty(displayName)) {
+            mDisplayNameTextView.setText(displayName);
         } else {
-            imageManager.cancelRequestAndClearImageView(mAvatarImageView);
-            mLoggedInAsHeading.setVisibility(View.GONE);
-            mUserDetailsCard.setVisibility(View.GONE);
+            mDisplayNameTextView.setText(username);
         }
     }
 
-    public void showSitesHeading(String text) {
+    void showSitesHeading(String text) {
         mMySitesHeadingTextView.setVisibility(View.VISIBLE);
         mMySitesHeadingTextView.setText(text);
     }
 
-    public void hideSitesHeading() {
+    void hideSitesHeading() {
         mMySitesHeadingTextView.setVisibility(View.GONE);
     }
 
+    private int getAvatarSize(Context context) {
+        return context.getResources().getDimensionPixelSize(R.dimen.avatar_sz_large);
+    }
+
     private String constructGravatarUrl(Context context, AccountModel account) {
-        int avatarSz = context.getResources().getDimensionPixelSize(R.dimen.avatar_sz_large);
-        return GravatarUtils.fixGravatarUrl(account.getAvatarUrl(), avatarSz);
+        return FakeGravatarUtils.fixGravatarUrl(account.getAvatarUrl(), getAvatarSize(context), STATUS_404);
+    }
+
+    private String constructGravatarUrl(Context context, SiteModel site) {
+        return FakeGravatarUtils.gravatarFromEmail(site.getEmail(), getAvatarSize(context), STATUS_404);
     }
 }
