@@ -1,7 +1,6 @@
 package org.wordpress.android.ui.main;
 
 import android.content.Context;
-import android.graphics.Typeface;
 import android.os.AsyncTask;
 import android.text.TextUtils;
 import android.util.TypedValue;
@@ -25,6 +24,7 @@ import org.wordpress.android.fluxc.store.SiteStore;
 import org.wordpress.android.ui.prefs.AppPrefs;
 import org.wordpress.android.util.AppLog;
 import org.wordpress.android.util.ContextExtensionsKt;
+import org.wordpress.android.util.DisplayUtils;
 import org.wordpress.android.util.SiteUtils;
 import org.wordpress.android.util.StringUtils;
 import org.wordpress.android.util.ViewUtilsKt;
@@ -57,10 +57,10 @@ public class SitePickerAdapter extends RecyclerView.Adapter<RecyclerView.ViewHol
         void onAfterLoad();
     }
 
-    public interface HeaderHandler {
-        RecyclerView.ViewHolder onCreateViewHolder(LayoutInflater layoutInflater, ViewGroup parent,
-                                                   boolean attachToRoot);
-        void onBindViewHolder(RecyclerView.ViewHolder holder, SiteList sites);
+    public interface ViewHolderHandler<T extends RecyclerView.ViewHolder> {
+        T onCreateViewHolder(LayoutInflater layoutInflater, ViewGroup parent, boolean attachToRoot);
+
+        void onBindViewHolder(T holder, SiteList sites);
     }
 
     /**
@@ -90,7 +90,8 @@ public class SitePickerAdapter extends RecyclerView.Adapter<RecyclerView.ViewHol
 
     private final LayoutInflater mInflater;
     private final HashSet<Integer> mSelectedPositions = new HashSet<>();
-    private final HeaderHandler mHeaderHandler;
+    private final ViewHolderHandler mHeaderHandler;
+    private final ViewHolderHandler mFooterHandler;
 
     private boolean mIsMultiSelectEnabled;
     private final boolean mIsInSearchMode;
@@ -114,6 +115,7 @@ public class SitePickerAdapter extends RecyclerView.Adapter<RecyclerView.ViewHol
 
     private static final int VIEW_TYPE_HEADER = 0;
     private static final int VIEW_TYPE_ITEM = 1;
+    private static final int VIEW_TYPE_FOOTER = 2;
 
     @Inject AccountStore mAccountStore;
     @Inject SiteStore mSiteStore;
@@ -149,7 +151,7 @@ public class SitePickerAdapter extends RecyclerView.Adapter<RecyclerView.ViewHol
                              SitePickerMode sitePickerMode
     ) {
         this(context, itemLayoutResourceId, currentLocalBlogId, lastSearch, isInSearchMode, dataLoadedListener,
-                null, null, sitePickerMode);
+                null, null, null, sitePickerMode);
     }
 
     public SitePickerAdapter(Context context,
@@ -158,11 +160,11 @@ public class SitePickerAdapter extends RecyclerView.Adapter<RecyclerView.ViewHol
                              String lastSearch,
                              boolean isInSearchMode,
                              OnDataLoadedListener dataLoadedListener,
-                             HeaderHandler headerHandler,
+                             ViewHolderHandler<?> headerHandler,
                              ArrayList<Integer> ignoreSitesIds
     ) {
         this(context, itemLayoutResourceId, currentLocalBlogId, lastSearch, isInSearchMode, dataLoadedListener,
-                headerHandler, ignoreSitesIds, SitePickerMode.DEFAULT_MODE);
+                headerHandler, null, ignoreSitesIds, SitePickerMode.DEFAULT_MODE);
     }
 
     public SitePickerAdapter(Context context,
@@ -171,7 +173,8 @@ public class SitePickerAdapter extends RecyclerView.Adapter<RecyclerView.ViewHol
                              String lastSearch,
                              boolean isInSearchMode,
                              OnDataLoadedListener dataLoadedListener,
-                             HeaderHandler headerHandler,
+                             ViewHolderHandler<?> headerHandler,
+                             ViewHolderHandler<?> footerHandler,
                              ArrayList<Integer> ignoreSitesIds,
                              SitePickerMode sitePickerMode
     ) {
@@ -199,6 +202,7 @@ public class SitePickerAdapter extends RecyclerView.Adapter<RecyclerView.ViewHol
                         context.getResources().getInteger(R.integer.selected_list_item_opacity));
 
         mHeaderHandler = headerHandler;
+        mFooterHandler = footerHandler;
         mSelectedItemPos = getPositionOffset();
 
         mIgnoreSitesIds = ignoreSitesIds;
@@ -210,7 +214,7 @@ public class SitePickerAdapter extends RecyclerView.Adapter<RecyclerView.ViewHol
 
     @Override
     public int getItemCount() {
-        return (mHeaderHandler != null ? 1 : 0) + mSites.size();
+        return (mHeaderHandler != null ? 1 : 0) + mSites.size() + (mFooterHandler != null ? 1 : 0);
     }
 
     private int getSitesCount() {
@@ -219,8 +223,11 @@ public class SitePickerAdapter extends RecyclerView.Adapter<RecyclerView.ViewHol
 
     @Override
     public long getItemId(int position) {
-        if (mHeaderHandler != null && position == 0) {
-            return RecyclerView.NO_ID;
+        int viewType = getItemViewType(position);
+        if (viewType == VIEW_TYPE_HEADER) {
+            return -1;
+        } else if (viewType == VIEW_TYPE_FOOTER) {
+            return -2;
         } else {
             return getItem(position).mLocalId;
         }
@@ -228,10 +235,12 @@ public class SitePickerAdapter extends RecyclerView.Adapter<RecyclerView.ViewHol
 
     @Override
     public int getItemViewType(int position) {
-        if (mHeaderHandler == null) {
-            return VIEW_TYPE_ITEM;
+        if (mHeaderHandler != null && position == 0) {
+            return VIEW_TYPE_HEADER;
+        } else if (mFooterHandler != null && position == getItemCount() - 1) {
+            return VIEW_TYPE_FOOTER;
         } else {
-            return position == 0 ? VIEW_TYPE_HEADER : VIEW_TYPE_ITEM;
+            return VIEW_TYPE_ITEM;
         }
     }
 
@@ -257,6 +266,8 @@ public class SitePickerAdapter extends RecyclerView.Adapter<RecyclerView.ViewHol
     public RecyclerView.ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         if (viewType == VIEW_TYPE_HEADER) {
             return mHeaderHandler.onCreateViewHolder(mInflater, parent, false);
+        } else if (viewType == VIEW_TYPE_FOOTER) {
+            return mFooterHandler.onCreateViewHolder(mInflater, parent, false);
         } else {
             View itemView = mInflater.inflate(mItemLayoutReourceId, parent, false);
             return new SiteViewHolder(itemView);
@@ -272,13 +283,18 @@ public class SitePickerAdapter extends RecyclerView.Adapter<RecyclerView.ViewHol
             return;
         }
 
+        if (viewType == VIEW_TYPE_FOOTER) {
+            mFooterHandler.onBindViewHolder(viewHolder, mSites);
+            return;
+        }
+
         SiteRecord site = getItem(position);
 
         final SiteViewHolder holder = (SiteViewHolder) viewHolder;
         holder.mTxtTitle.setText(site.getBlogNameOrHomeURL());
         holder.mTxtDomain.setText(site.mHomeURL);
-        mImageManager.load(holder.mImgBlavatar, ImageType.BLAVATAR, site.mBlavatarUrl);
-
+        mImageManager.loadImageWithCorners(holder.mImgBlavatar, ImageType.BLAVATAR_ROUNDED_CORNERS, site.mBlavatarUrl,
+                DisplayUtils.dpToPx(holder.itemView.getContext(), 4));
 
         if ((site.mLocalId == mCurrentLocalId && !mIsMultiSelectEnabled
              && mSitePickerMode == SitePickerMode.DEFAULT_MODE)
@@ -293,8 +309,6 @@ public class SitePickerAdapter extends RecyclerView.Adapter<RecyclerView.ViewHol
         if (holder.mIsSiteHidden == null || holder.mIsSiteHidden != site.mIsHidden) {
             holder.mIsSiteHidden = site.mIsHidden;
             holder.mTxtTitle.setAlpha(site.mIsHidden ? mDisabledSiteOpacity : 1f);
-            holder.mTxtTitle
-                    .setTypeface(holder.mTxtTitle.getTypeface(), site.mIsHidden ? Typeface.NORMAL : Typeface.BOLD);
             holder.mImgBlavatar.setAlpha(site.mIsHidden ? mDisabledSiteOpacity : 1f);
         }
 
@@ -389,7 +403,7 @@ public class SitePickerAdapter extends RecyclerView.Adapter<RecyclerView.ViewHol
 
     public int getItemPosByLocalId(int localId) {
         int positionInSitesArray = mSites.indexOfSiteId(localId);
-        
+
         return mSites.size() != 0 && positionInSitesArray > -1 ? positionInSitesArray : -1;
     }
 
