@@ -334,6 +334,9 @@ public final class AnalyticsTracker {
         CONNECT_JETPACK_FAILED,
         PUSH_NOTIFICATION_RECEIVED,
         PUSH_NOTIFICATION_TAPPED, // Same of opened
+        UNIFIED_LOGIN_STEP,
+        UNIFIED_LOGIN_INTERACTION,
+        UNIFIED_LOGIN_FAILURE,
         LOGIN_ACCESSED,
         LOGIN_MAGIC_LINK_EXITED,
         LOGIN_MAGIC_LINK_FAILED,
@@ -346,11 +349,6 @@ public final class AnalyticsTracker {
         LOGIN_AUTOFILL_CREDENTIALS_FILLED,
         LOGIN_AUTOFILL_CREDENTIALS_UPDATED,
         LOGIN_PROLOGUE_PAGED,
-        LOGIN_PROLOGUE_PAGED_JETPACK,
-        LOGIN_PROLOGUE_PAGED_NOTIFICATIONS,
-        LOGIN_PROLOGUE_PAGED_POST,
-        LOGIN_PROLOGUE_PAGED_READER,
-        LOGIN_PROLOGUE_PAGED_STATS,
         LOGIN_PROLOGUE_VIEWED,
         LOGIN_EMAIL_FORM_VIEWED,
         LOGIN_MAGIC_LINK_OPEN_EMAIL_CLIENT_VIEWED,
