@@ -1188,6 +1188,12 @@ public class AnalyticsTrackerNosara extends Tracker {
                 return "push_notification_received";
             case PUSH_NOTIFICATION_TAPPED:
                 return "push_notification_alert_tapped";
+            case UNIFIED_LOGIN_STEP:
+                return "unified_login_step";
+            case UNIFIED_LOGIN_INTERACTION:
+                return "unified_login_interaction";
+            case UNIFIED_LOGIN_FAILURE:
+                return "unified_login_failure";
             case LOGIN_ACCESSED:
                 return "login_accessed";
             case LOGIN_MAGIC_LINK_EXITED:
@@ -1212,16 +1218,6 @@ public class AnalyticsTrackerNosara extends Tracker {
                 return "login_autofill_credentials_updated";
             case LOGIN_PROLOGUE_PAGED:
                 return "login_prologue_paged";
-            case LOGIN_PROLOGUE_PAGED_JETPACK:
-                return "login_prologue_paged_jetpack";
-            case LOGIN_PROLOGUE_PAGED_NOTIFICATIONS:
-                return "login_prologue_paged_notifications";
-            case LOGIN_PROLOGUE_PAGED_POST:
-                return "login_prologue_paged_post";
-            case LOGIN_PROLOGUE_PAGED_READER:
-                return "login_prologue_paged_reader";
-            case LOGIN_PROLOGUE_PAGED_STATS:
-                return "login_prologue_paged_stats";
             case LOGIN_PROLOGUE_VIEWED:
                 return "login_prologue_viewed";
             case LOGIN_EMAIL_FORM_VIEWED:
