@@ -18,7 +18,9 @@ import androidx.annotation.LayoutRes;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.annotation.StringRes;
+import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.widget.Toolbar;
 
 import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 
@@ -89,7 +91,7 @@ public class Login2FaFragment extends LoginBaseFormFragment<LoginListener> imple
 
     ArrayList<Integer> mOldSitesIDs;
 
-    private Button mSecondaryButton;
+    private Button mOtpButton;
     private String mEmailAddress;
     private String mIdToken;
     private String mNonce;
@@ -172,27 +174,33 @@ public class Login2FaFragment extends LoginBaseFormFragment<LoginListener> imple
 
         // restrict the allowed input chars to just numbers
         m2FaInput.getEditText().setKeyListener(DigitsKeyListener.getInstance("0123456789"));
-    }
 
-    @Override
-    protected void setupBottomButtons(Button secondaryButton, Button primaryButton) {
-        secondaryButton.setText(R.string.login_text_otp);
-        secondaryButton.setOnClickListener(new OnClickListener() {
+        mOtpButton = rootView.findViewById(R.id.login_otp_button);
+        mOtpButton.setText(mSentSmsCode ? R.string.login_text_otp_another : R.string.login_text_otp);
+        mOtpButton.setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(View v) {
                 if (isAdded()) {
+                    mAnalyticsListener.trackSendCodeWithTextClicked();
                     doAuthAction(R.string.requesting_otp, "", true);
                 }
             }
         });
-        secondaryButton.setText(getString(mSentSmsCode ? R.string.login_text_otp_another : R.string.login_text_otp));
-        mSecondaryButton = secondaryButton;
+    }
 
+    @Override
+    protected void setupBottomButtons(Button secondaryButton, Button primaryButton) {
+        secondaryButton.setVisibility(View.GONE);
         primaryButton.setOnClickListener(new OnClickListener() {
             public void onClick(View v) {
                 next();
             }
         });
+    }
+
+    @Override
+    protected void buildToolbar(Toolbar toolbar, ActionBar actionBar) {
+        actionBar.setTitle(R.string.log_in);
     }
 
     @Override
@@ -278,6 +286,7 @@ public class Login2FaFragment extends LoginBaseFormFragment<LoginListener> imple
     }
 
     protected void next() {
+        mAnalyticsListener.trackSubmit2faCodeClicked();
         if (TextUtils.isEmpty(m2FaInput.getEditText().getText())) {
             show2FaError(getString(R.string.login_empty_2fa));
             return;
@@ -368,6 +377,7 @@ public class Login2FaFragment extends LoginBaseFormFragment<LoginListener> imple
     }
 
     private void show2FaError(String message) {
+        mAnalyticsListener.trackFailure(message);
         m2FaInput.setError(message);
     }
 
@@ -389,7 +399,7 @@ public class Login2FaFragment extends LoginBaseFormFragment<LoginListener> imple
                 // TODO: FluxC: could be specific?
             default:
                 AppLog.e(T.NUX, "Server response: " + errorMessage);
-
+                mAnalyticsListener.trackFailure(errorMessage);
                 ToastUtils.showToast(getActivity(),
                         errorMessage == null ? getString(R.string.error_generic) : errorMessage);
                 break;
@@ -397,6 +407,7 @@ public class Login2FaFragment extends LoginBaseFormFragment<LoginListener> imple
     }
 
     private void showErrorDialog(String message) {
+        mAnalyticsListener.trackFailure(message);
         AlertDialog dialog = new MaterialAlertDialogBuilder(getActivity())
                 .setMessage(message)
                 .setPositiveButton(R.string.login_error_button, null)
@@ -514,7 +525,7 @@ public class Login2FaFragment extends LoginBaseFormFragment<LoginListener> imple
 
     private void setTextForSms() {
         mLabel.setText(getString(R.string.enter_verification_code_sms, mPhoneNumber));
-        mSecondaryButton.setText(getString(R.string.login_text_otp_another));
+        mOtpButton.setText(getString(R.string.login_text_otp_another));
         mSentSmsCode = true;
     }
 }
