@@ -8,6 +8,7 @@ import android.content.IntentSender;
 import android.os.Bundle;
 import android.text.Editable;
 import android.text.Html;
+import android.text.Spanned;
 import android.text.TextWatcher;
 import android.util.Patterns;
 import android.view.View;
@@ -21,7 +22,9 @@ import android.widget.TextView;
 import androidx.annotation.LayoutRes;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.widget.Toolbar;
 
 import com.google.android.gms.auth.api.Auth;
 import com.google.android.gms.auth.api.credentials.Credential;
@@ -39,6 +42,8 @@ import org.greenrobot.eventbus.ThreadMode;
 import org.wordpress.android.fluxc.generated.AccountActionBuilder;
 import org.wordpress.android.fluxc.store.AccountStore;
 import org.wordpress.android.fluxc.store.AccountStore.OnAvailabilityChecked;
+import org.wordpress.android.login.SignupBottomSheetDialogFragment.SignupSheetListener;
+import org.wordpress.android.login.util.ContextExtensionsKt;
 import org.wordpress.android.login.util.SiteUtils;
 import org.wordpress.android.login.widgets.WPLoginInputRow;
 import org.wordpress.android.login.widgets.WPLoginInputRow.OnEditorCommitListener;
@@ -46,6 +51,7 @@ import org.wordpress.android.util.ActivityUtils;
 import org.wordpress.android.util.AppLog;
 import org.wordpress.android.util.AppLog.T;
 import org.wordpress.android.util.EditTextUtils;
+import org.wordpress.android.util.HtmlUtils;
 import org.wordpress.android.util.NetworkUtils;
 
 import java.util.ArrayList;
@@ -70,6 +76,9 @@ public class LoginEmailFragment extends LoginBaseFormFragment<LoginListener> imp
     private static final int EMAIL_CREDENTIALS_REQUEST_CODE = 25100;
 
     private static final String ARG_LOGIN_SITE_URL = "ARG_LOGIN_SITE_URL";
+    private static final String ARG_SIGNUP_FROM_LOGIN_ENABLED = "ARG_SIGNUP_FROM_LOGIN_ENABLED";
+    private static final String ARG_SITE_LOGIN_ENABLED = "ARG_SITE_LOGIN_ENABLED";
+    private static final String ARG_SHOULD_USE_NEW_LAYOUT = "ARG_SHOULD_USE_NEW_LAYOUT";
 
     public static final String TAG = "login_email_fragment_tag";
     public static final int MAX_EMAIL_LENGTH = 100;
@@ -80,6 +89,9 @@ public class LoginEmailFragment extends LoginBaseFormFragment<LoginListener> imp
     private String mRequestedEmail;
     private boolean mIsSocialLogin;
     private Integer mCurrentEmailErrorRes = null;
+    private boolean mIsSignupFromLoginEnabled;
+    private boolean mIsSiteLoginEnabled;
+    private boolean mShouldUseNewLayout;
 
     protected WPLoginInputRow mEmailInput;
     protected boolean mHasDismissedEmailHints;
@@ -94,9 +106,20 @@ public class LoginEmailFragment extends LoginBaseFormFragment<LoginListener> imp
         return fragment;
     }
 
+    public static LoginEmailFragment newInstance(boolean isSignupFromLoginEnabled, boolean isSiteLoginEnabled,
+                                                 boolean shouldUseNewLayout) {
+        LoginEmailFragment fragment = new LoginEmailFragment();
+        Bundle args = new Bundle();
+        args.putBoolean(ARG_SIGNUP_FROM_LOGIN_ENABLED, isSignupFromLoginEnabled);
+        args.putBoolean(ARG_SITE_LOGIN_ENABLED, isSiteLoginEnabled);
+        args.putBoolean(ARG_SHOULD_USE_NEW_LAYOUT, shouldUseNewLayout);
+        fragment.setArguments(args);
+        return fragment;
+    }
+
     @Override
     protected @LayoutRes int getContentLayout() {
-        return R.layout.login_email_screen;
+        return mShouldUseNewLayout ? R.layout.login_email_screen : R.layout.login_email_screen_old;
     }
 
     @Override
@@ -115,7 +138,11 @@ public class LoginEmailFragment extends LoginBaseFormFragment<LoginListener> imp
                 break;
             case FULL:
             case WPCOM_LOGIN_ONLY:
-                label.setText(R.string.enter_email_wordpress_com);
+                if (mShouldUseNewLayout) {
+                    label.setText(R.string.enter_email_to_continue_wordpress_com);
+                } else {
+                    label.setText(R.string.enter_email_wordpress_com);
+                }
                 break;
             case WOO_LOGIN_MODE:
                 label.setText(getString(R.string.enter_email_for_site, mLoginSiteUrl));
@@ -133,7 +160,24 @@ public class LoginEmailFragment extends LoginBaseFormFragment<LoginListener> imp
     protected void setupContent(ViewGroup rootView) {
         // important for accessibility - talkback
         getActivity().setTitle(R.string.email_address_login_title);
-        mEmailInput = rootView.findViewById(R.id.login_email_row);
+
+        setupEmailInput((WPLoginInputRow) rootView.findViewById(R.id.login_email_row));
+
+        if (mShouldUseNewLayout) {
+            setupContinueButton((Button) rootView.findViewById(R.id.login_continue_button));
+            setupTosButtons(
+                    (Button) rootView.findViewById(R.id.continue_tos),
+                    (Button) rootView.findViewById(R.id.continue_with_google_tos));
+            setupSocialButtons((Button) rootView.findViewById(R.id.continue_with_google));
+        } else {
+            setupAlternativeButtons(
+                    (LinearLayout) rootView.findViewById(R.id.login_google_button),
+                    (LinearLayout) rootView.findViewById(R.id.login_site_button));
+        }
+    }
+
+    private void setupEmailInput(WPLoginInputRow emailInput) {
+        mEmailInput = emailInput;
         if (BuildConfig.DEBUG) {
             mEmailInput.getEditText().setText(BuildConfig.DEBUG_WPCOM_LOGIN_EMAIL);
         }
@@ -148,6 +192,7 @@ public class LoginEmailFragment extends LoginBaseFormFragment<LoginListener> imp
             @Override
             public void onFocusChange(View view, boolean hasFocus) {
                 if (hasFocus && !mIsDisplayingEmailHints && !mHasDismissedEmailHints) {
+                    mAnalyticsListener.trackSelectEmailField();
                     mIsDisplayingEmailHints = true;
                     getEmailHints();
                 }
@@ -156,34 +201,67 @@ public class LoginEmailFragment extends LoginBaseFormFragment<LoginListener> imp
         mEmailInput.getEditText().setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(View view) {
+                mAnalyticsListener.trackSelectEmailField();
                 if (!mIsDisplayingEmailHints && !mHasDismissedEmailHints) {
+                    mAnalyticsListener.trackSelectEmailField();
                     mIsDisplayingEmailHints = true;
                     getEmailHints();
                 }
             }
         });
+    }
 
-        LinearLayout googleLoginButton = rootView.findViewById(R.id.login_google_button);
+    private void setupContinueButton(Button continueButton) {
+        continueButton.setOnClickListener(new OnClickListener() {
+            public void onClick(View view) {
+                onContinueClicked();
+            }
+        });
+    }
+
+    private void updateContinueButtonEnabledStatus() {
+        View view = getView();
+        if (view != null) {
+            Button continueButton = (Button) view.findViewById(R.id.login_continue_button);
+            String currentEmail = mEmailInput.getEditText().getText().toString();
+            continueButton.setEnabled(!currentEmail.trim().isEmpty());
+        }
+    }
+
+    private void setupTosButtons(Button continueTosButton, Button continueWithGoogleTosButton) {
+        OnClickListener onClickListener = new OnClickListener() {
+            public void onClick(View view) {
+                Context context = getContext();
+                if ((context instanceof SignupSheetListener)) {
+                    ((SignupSheetListener) context).onSignupSheetTermsOfServiceClicked();
+                }
+            }
+        };
+
+        continueTosButton.setOnClickListener(onClickListener);
+        continueTosButton.setText(formatTosText(R.string.continue_terms_of_service_text));
+
+        continueWithGoogleTosButton.setOnClickListener(onClickListener);
+        continueWithGoogleTosButton.setText(formatTosText(R.string.continue_with_google_terms_of_service_text));
+    }
+
+    private void setupSocialButtons(Button continueWithGoogleButton) {
+        continueWithGoogleButton.setOnClickListener(new OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                onGoogleSigninClicked();
+            }
+        });
+    }
+
+    private void setupAlternativeButtons(LinearLayout googleLoginButton, LinearLayout siteLoginButton) {
         googleLoginButton.setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(View view) {
-                mAnalyticsListener.trackSocialButtonClick();
-                ActivityUtils.hideKeyboardForced(mEmailInput.getEditText());
-
-                if (NetworkUtils.checkConnection(getActivity())) {
-                    if (isAdded()) {
-                        mOldSitesIDs = SiteUtils.getCurrentSiteIds(mSiteStore, false);
-                        mIsSocialLogin = true;
-                        mLoginListener.addGoogleLoginFragment();
-                    } else {
-                        AppLog.e(T.NUX, "Google login could not be started.  LoginEmailFragment was not attached.");
-                        showErrorDialog(getString(R.string.login_error_generic_start));
-                    }
-                }
+                onGoogleSigninClicked();
             }
         });
 
-        LinearLayout siteLoginButton = rootView.findViewById(R.id.login_site_button);
         siteLoginButton.setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(View view) {
@@ -200,8 +278,8 @@ public class LoginEmailFragment extends LoginBaseFormFragment<LoginListener> imp
             }
         });
 
-        ImageView siteLoginButtonIcon = rootView.findViewById(R.id.login_site_button_icon);
-        TextView siteLoginButtonText = rootView.findViewById(R.id.login_site_button_text);
+        ImageView siteLoginButtonIcon = siteLoginButton.findViewById(R.id.login_site_button_icon);
+        TextView siteLoginButtonText = siteLoginButton.findViewById(R.id.login_site_button_text);
 
         switch (mLoginListener.getLoginMode()) {
             case WOO_LOGIN_MODE:
@@ -211,6 +289,7 @@ public class LoginEmailFragment extends LoginBaseFormFragment<LoginListener> imp
             case FULL:
             case WPCOM_LOGIN_ONLY:
             case SHARE_INTENT:
+                siteLoginButton.setVisibility(mIsSiteLoginEnabled ? View.VISIBLE : View.GONE);
                 siteLoginButtonIcon.setImageResource(R.drawable.ic_domains_grey_24dp);
                 siteLoginButtonText.setText(R.string.enter_site_address_instead);
                 break;
@@ -227,9 +306,19 @@ public class LoginEmailFragment extends LoginBaseFormFragment<LoginListener> imp
 
     @Override
     protected void setupBottomButtons(Button secondaryButton, Button primaryButton) {
-        if (mLoginListener.getLoginMode() == LoginMode.JETPACK_STATS) {
-            secondaryButton.setText(Html.fromHtml(String.format(getResources().getString(
-                    R.string.login_email_button_signup), "<u>", "</u>")));
+        if (mShouldUseNewLayout) {
+            secondaryButton.setVisibility(View.GONE);
+            primaryButton.setVisibility(View.GONE);
+        } else {
+            setupSecondaryButton(secondaryButton);
+            setupPrimaryButton(primaryButton);
+        }
+    }
+
+    private void setupSecondaryButton(Button secondaryButton) {
+        // Show Sign-Up button if login mode is Jetpack and signup from login is not enabled
+        if (mLoginListener.getLoginMode() == LoginMode.JETPACK_STATS && !mIsSignupFromLoginEnabled) {
+            secondaryButton.setText(formatUnderlinedText(R.string.login_email_button_signup));
             secondaryButton.setOnClickListener(new OnClickListener() {
                 public void onClick(View view) {
                     mLoginListener.doStartSignup();
@@ -250,12 +339,44 @@ public class LoginEmailFragment extends LoginBaseFormFragment<LoginListener> imp
         } else {
             secondaryButton.setVisibility(View.GONE);
         }
+    }
 
+    private void setupPrimaryButton(Button primaryButton) {
         primaryButton.setOnClickListener(new OnClickListener() {
             public void onClick(View view) {
-                next(getCleanedEmail());
+                onContinueClicked();
             }
         });
+    }
+
+    private Spanned formatTosText(int stringResId) {
+        final int primaryColorResId = ContextExtensionsKt.getColorResIdFromAttribute(getContext(), R.attr.colorPrimary);
+        final String primaryColorHtml = HtmlUtils.colorResToHtmlColor(getContext(), primaryColorResId);
+        return Html.fromHtml(getString(stringResId, "<u><font color='" + primaryColorHtml + "'>", "</font></u>"));
+    }
+
+    private Spanned formatUnderlinedText(int stringResId) {
+        return Html.fromHtml(getString(stringResId, "<u>", "</u>"));
+    }
+
+    private void onContinueClicked() {
+        next(getCleanedEmail());
+    }
+
+    private void onGoogleSigninClicked() {
+        mAnalyticsListener.trackSocialButtonClick();
+        ActivityUtils.hideKeyboardForced(mEmailInput.getEditText());
+
+        if (NetworkUtils.checkConnection(getActivity())) {
+            if (isAdded()) {
+                mOldSitesIDs = SiteUtils.getCurrentSiteIds(mSiteStore, false);
+                mIsSocialLogin = true;
+                mLoginListener.addGoogleLoginFragment();
+            } else {
+                AppLog.e(T.NUX, "Google login could not be started.  LoginEmailFragment was not attached.");
+                showErrorDialog(getString(R.string.login_error_generic_start));
+            }
+        }
     }
 
     @Override
@@ -284,6 +405,9 @@ public class LoginEmailFragment extends LoginBaseFormFragment<LoginListener> imp
         Bundle args = getArguments();
         if (args != null) {
             mLoginSiteUrl = args.getString(ARG_LOGIN_SITE_URL, "");
+            mIsSignupFromLoginEnabled = args.getBoolean(ARG_SIGNUP_FROM_LOGIN_ENABLED, false);
+            mIsSiteLoginEnabled = args.getBoolean(ARG_SITE_LOGIN_ENABLED, true);
+            mShouldUseNewLayout = args.getBoolean(ARG_SHOULD_USE_NEW_LAYOUT, false);
         }
     }
 
@@ -329,6 +453,13 @@ public class LoginEmailFragment extends LoginBaseFormFragment<LoginListener> imp
     }
 
     @Override
+    public void onResume() {
+        super.onResume();
+        mAnalyticsListener.emailFormScreenResumed();
+        updateContinueButtonEnabledStatus();
+    }
+
+    @Override
     public void onSaveInstanceState(Bundle outState) {
         super.onSaveInstanceState(outState);
         outState.putIntegerArrayList(KEY_OLD_SITES_IDS, mOldSitesIDs);
@@ -342,7 +473,17 @@ public class LoginEmailFragment extends LoginBaseFormFragment<LoginListener> imp
         }
     }
 
+    @Override
+    protected void buildToolbar(Toolbar toolbar, ActionBar actionBar) {
+        if (mShouldUseNewLayout) {
+            actionBar.setTitle(R.string.get_started);
+        } else {
+            super.buildToolbar(toolbar, actionBar);
+        }
+    }
+
     protected void next(String email) {
+        mAnalyticsListener.trackSubmitClicked();
         if (!NetworkUtils.checkConnection(getActivity())) {
             return;
         }
@@ -367,7 +508,7 @@ public class LoginEmailFragment extends LoginBaseFormFragment<LoginListener> imp
 
     private void showEmailError() {
         if (mCurrentEmailErrorRes != null) {
-             showEmailError(mCurrentEmailErrorRes);
+            showEmailError(mCurrentEmailErrorRes);
         }
     }
 
@@ -406,11 +547,14 @@ public class LoginEmailFragment extends LoginBaseFormFragment<LoginListener> imp
         mEmailInput.setError(null);
         mIsSocialLogin = false;
         clearEmailError();
+        updateContinueButtonEnabledStatus();
     }
 
     private void showEmailError(int messageId) {
         mCurrentEmailErrorRes = messageId;
-        mEmailInput.setError(getString(messageId));
+        String errorMessage = getString(messageId);
+        mAnalyticsListener.trackFailure(errorMessage);
+        mEmailInput.setError(errorMessage);
     }
 
     private void showErrorDialog(String message) {
@@ -458,9 +602,16 @@ public class LoginEmailFragment extends LoginBaseFormFragment<LoginListener> imp
         switch (event.type) {
             case EMAIL:
                 if (event.isAvailable) {
-                    // email address is available on wpcom, so apparently the user can't login with that one.
                     ActivityUtils.hideKeyboardForced(mEmailInput);
-                    showEmailError(R.string.email_not_registered_wpcom);
+                    if (mIsSignupFromLoginEnabled) {
+                        if (mLoginListener != null) {
+                            mLoginListener.gotUnregisteredEmail(event.value);
+                        }
+                    } else {
+                        // email address is available on wpcom, so apparently the user can't login with that one.
+                        mAnalyticsListener.trackFailure("Email not registered WP.com");
+                        showEmailError(R.string.email_not_registered_wpcom);
+                    }
                 } else if (mLoginListener != null) {
                     ActivityUtils.hideKeyboardForced(mEmailInput);
                     mLoginListener.gotWpcomEmail(event.value, false);
