@@ -3,7 +3,6 @@ package org.wordpress.android.login;
 import android.app.ProgressDialog;
 import android.content.Context;
 import android.content.DialogInterface;
-import android.graphics.drawable.Drawable;
 import android.os.Bundle;
 import android.text.Html;
 import android.view.LayoutInflater;
@@ -23,13 +22,6 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 import androidx.fragment.app.Fragment;
 
-import com.bumptech.glide.Glide;
-import com.bumptech.glide.load.DataSource;
-import com.bumptech.glide.load.engine.GlideException;
-import com.bumptech.glide.request.RequestListener;
-import com.bumptech.glide.request.RequestOptions;
-import com.bumptech.glide.request.target.Target;
-
 import org.greenrobot.eventbus.Subscribe;
 import org.greenrobot.eventbus.ThreadMode;
 import org.wordpress.android.fluxc.Dispatcher;
@@ -39,8 +31,9 @@ import org.wordpress.android.fluxc.store.AccountStore.AuthEmailPayload;
 import org.wordpress.android.fluxc.store.AccountStore.AuthEmailPayloadScheme;
 import org.wordpress.android.fluxc.store.AccountStore.AuthEmailPayloadSource;
 import org.wordpress.android.fluxc.store.AccountStore.OnAuthEmailSent;
+import org.wordpress.android.login.util.AvatarHelper;
+import org.wordpress.android.login.util.AvatarHelper.AvatarRequestListener;
 import org.wordpress.android.util.AppLog;
-import org.wordpress.android.util.GravatarUtils;
 import org.wordpress.android.util.NetworkUtils;
 import org.wordpress.android.util.ToastUtils;
 
@@ -126,6 +119,7 @@ public class LoginMagicLinkRequestFragment extends Fragment {
         mRequestMagicLinkButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+                mAnalyticsListener.trackRequestMagicLinkClick();
                 if (mLoginListener != null) {
                     if (NetworkUtils.checkConnection(getActivity())) {
                         showMagicLinkRequestProgressDialog();
@@ -142,6 +136,7 @@ public class LoginMagicLinkRequestFragment extends Fragment {
         view.findViewById(R.id.login_enter_password).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+                mAnalyticsListener.trackLoginWithPasswordClick();
                 if (mLoginListener != null) {
                     mLoginListener.usePasswordInstead(mEmail);
                 }
@@ -150,6 +145,9 @@ public class LoginMagicLinkRequestFragment extends Fragment {
 
         mAvatarProgressBar = view.findViewById(R.id.avatar_progress);
         ImageView avatarView = view.findViewById(R.id.gravatar);
+
+        TextView emailView = view.findViewById(R.id.email);
+        emailView.setText(mEmail);
 
         // Design changes added to the Woo Magic link sign-in
         if (mVerifyMagicLinkEmail) {
@@ -169,28 +167,11 @@ public class LoginMagicLinkRequestFragment extends Fragment {
 
             mRequestMagicLinkButton.setText(getString(R.string.send_verification_email));
         } else {
-            Glide.with(this)
-                 .load(GravatarUtils.gravatarFromEmail(mEmail,
-                         getContext().getResources().getDimensionPixelSize(R.dimen.avatar_sz_login)))
-                 .apply(RequestOptions.circleCropTransform())
-                 .apply(RequestOptions.placeholderOf(R.drawable.ic_gridicons_user_circle_100dp))
-                 .apply(RequestOptions.errorOf(R.drawable.ic_gridicons_user_circle_100dp))
-                 .listener(new RequestListener<Drawable>() {
-                     @Override
-                     public boolean onLoadFailed(@Nullable GlideException e, Object o, Target<Drawable> target,
-                                                 boolean b) {
-                         mAvatarProgressBar.setVisibility(View.GONE);
-                         return false;
-                     }
-
-                     @Override
-                     public boolean onResourceReady(Drawable drawable, Object o, Target<Drawable> target,
-                                                    DataSource dataSource, boolean b) {
-                         mAvatarProgressBar.setVisibility(View.GONE);
-                         return false;
-                     }
-                 })
-                 .into(avatarView);
+            AvatarHelper.loadAvatarFromEmail(this, mEmail, avatarView, new AvatarRequestListener() {
+                @Override public void onRequestFinished() {
+                    mAvatarProgressBar.setVisibility(View.GONE);
+                }
+            });
         }
 
         return view;
@@ -205,7 +186,7 @@ public class LoginMagicLinkRequestFragment extends Fragment {
 
         ActionBar actionBar = ((AppCompatActivity) getActivity()).getSupportActionBar();
         if (actionBar != null) {
-            actionBar.setDisplayShowTitleEnabled(false);
+            actionBar.setTitle(R.string.log_in);
             actionBar.setDisplayHomeAsUpEnabled(true);
         }
 
@@ -249,6 +230,7 @@ public class LoginMagicLinkRequestFragment extends Fragment {
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         if (item.getItemId() == R.id.help) {
+            mAnalyticsListener.trackShowHelpClick();
             if (mLoginListener != null) {
                 mLoginListener.helpMagicLinkRequest(mEmail);
             }
@@ -330,6 +312,7 @@ public class LoginMagicLinkRequestFragment extends Fragment {
             HashMap<String, String> errorProperties = new HashMap<>();
             errorProperties.put(ERROR_KEY, event.error.message);
             mAnalyticsListener.trackMagicLinkFailed(errorProperties);
+            mAnalyticsListener.trackFailure(event.error.message);
 
             AppLog.e(AppLog.T.API, "OnAuthEmailSent has error: " + event.error.type + " - " + event.error.message);
             if (isAdded()) {
