@@ -80,7 +80,7 @@ public class SignupMagicLinkFragment extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View layout = inflater.inflate(R.layout.signup_magic_link, container, false);
+        View layout = inflater.inflate(R.layout.signup_magic_link_screen, container, false);
 
         mOpenMailButton = layout.findViewById(R.id.signup_magic_link_button);
         mOpenMailButton.setOnClickListener(new View.OnClickListener() {
@@ -113,12 +113,12 @@ public class SignupMagicLinkFragment extends Fragment {
         ActionBar actionBar = ((AppCompatActivity) getActivity()).getSupportActionBar();
 
         if (actionBar != null) {
-            actionBar.setDisplayShowTitleEnabled(false);
+            actionBar.setTitle(R.string.sign_up_label);
             actionBar.setDisplayHomeAsUpEnabled(true);
         }
 
         if (savedInstanceState == null) {
-            mAnalyticsListener.trackMagicLinkOpenEmailClientViewed();
+            mAnalyticsListener.trackSignupMagicLinkOpenEmailClientViewed();
         }
     }
 
@@ -157,6 +157,7 @@ public class SignupMagicLinkFragment extends Fragment {
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         if (item.getItemId() == R.id.help) {
+            mAnalyticsListener.trackShowHelpClick();
             if (mLoginListener != null) {
                 mLoginListener.helpSignupMagicLinkScreen(mEmail);
             }
@@ -220,6 +221,7 @@ public class SignupMagicLinkFragment extends Fragment {
     }
 
     protected void showErrorDialog(String message) {
+        mAnalyticsListener.trackFailure(message);
         DialogInterface.OnClickListener dialogListener = new DialogInterface.OnClickListener() {
             @Override
             public void onClick(DialogInterface dialog, int which) {
